+++ common/src/main/java/com/faboslav/structurify/common/config/StructurifyConfigSerializer.java
@@ -25,14 +25,38 @@ public final class StructurifyConfigSerializer
 			Structurify.getLogger().info("Disabled {} structures: {}", disabledStructures.size(), disabledStructures);
 		}
 
-		List<String> changedStructureSets = Structurify.getConfig().getStructureSetData().entrySet()
+		List<String> changedStructures = Structurify.getConfig().getStructureData().entrySet()
 			.stream()
-			.filter(entry -> entry.getValue().isUsingDefaultValues())
+			.filter(entry -> !entry.getValue().isUsingDefaultValues())
 			.map(Map.Entry::getKey)
 			.toList();
 
+		List<String> changedStructureNamespaces = Structurify.getConfig().getStructureData().entrySet()
+			.stream()
+			.filter(entry -> !entry.getValue().isUsingDefaultValues())
+			.map(Map.Entry::getKey)
+			.toList();
+
+		List<String> changedStructureSets = Structurify.getConfig().getStructureSetData().entrySet()
+			.stream()
+			.filter(entry -> !entry.getValue().isUsingDefaultValues())
+			.map(Map.Entry::getKey)
+			.toList();
+
+		if (Structurify.getConfig().preventStructureOverlap) {
+			Structurify.getLogger().info("Enabled structure overlap prevention");
+		}
+
 		if (Structurify.getConfig().enableGlobalSpacingAndSeparationModifier && Structurify.getConfig().globalSpacingAndSeparationModifier != 1.0D) {
 			Structurify.getLogger().info("Enabled global spacing and separation modifier with value of {}", Structurify.getConfig().globalSpacingAndSeparationModifier);
+		}
+
+		if (!changedStructures.isEmpty()) {
+			Structurify.getLogger().info("Changed settings of {} structures: {}", changedStructures.size(), changedStructures);
+		}
+
+		if (!changedStructureNamespaces.isEmpty()) {
+			Structurify.getLogger().info("Changed settings of {} structure namespaces: {}", changedStructureNamespaces.size(), changedStructureNamespaces);
 		}
 
 		if (!changedStructureSets.isEmpty()) {