+++ common/src/main/java/com/faboslav/structurify/common/config/data/StructureLikeData.java
@@ -3,6 +3,7 @@ package com.faboslav.structurify.common.config.data;
 import com.faboslav.structurify.common.config.data.structure.BiomeCheckData;
 import com.faboslav.structurify.common.config.data.structure.DistanceFromWorldCenterCheckData;
 import com.faboslav.structurify.common.config.data.structure.FlatnessCheckData;
+import com.faboslav.structurify.common.config.data.structure.OverlapCheckData;
 
 public interface StructureLikeData
 {
@@ -15,6 +16,10 @@ public interface StructureLikeData
 	DistanceFromWorldCenterCheckData getDistanceFromWorldCenterCheckData();
 
 	void setDistanceFromWorldCenterCheckData(DistanceFromWorldCenterCheckData distanceFromWorldCenterCheckData);
+
+	OverlapCheckData getOverlapCheckData();
+
+	void setOverlapCheckData(OverlapCheckData overlapCheckData);
 
 	FlatnessCheckData getFlatnessCheckData();